--- src/application/services/professionals.service.ts
@@ -1,11 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { UserRepository } from '@/database/repositories/user.repository';
-
-@Injectable()
-export class ProfessionalsService {
-	constructor(private readonly userRepository: UserRepository) {}
-
-	async listProfessionals() {
-		return await this.userRepository.listProfessionals();
-	}
-}
