--- src/application/services/treatment-type.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { TreatmentTypeService } from '@services/treatment-type.service';
-
-describe('TreatmentTypeService', () => {
-	let service: TreatmentTypeService;
-
-	beforeEach(async () => {
-		const module: TestingModule = await Test.createTestingModule({
-			providers: [TreatmentTypeService],
-		}).compile();
-
-		service = module.get<TreatmentTypeService>(TreatmentTypeService);
-	});
-
-	it('should be defined', () => {
-		expect(service).toBeDefined();
-	});
-});
