+++ src/domain/services/professionals/professionals.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+import { UserRepository } from '@/infra/repositories/user.repository';
+
+@Injectable()
+export class ProfessionalsService {
+  constructor(private readonly userRepository: UserRepository) {}
+
+  async listProfessionals() {
+    return await this.userRepository.listProfessionals();
+  }
+}
